backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
pydantic==2.11.5
PyJWT==2.10.1
passlib==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
